Attempt to fix leftover keyboard
Often on leaving the app the keyboard would be left showing.  I think
this was a problem with PlayActivity trying to hide the keyboard from
the clue TextView.

Fixed by setting focus to the board view (requesting it in layout
play.xml) and hiding the keyboard from getCurrentFocus on pause.

Similarly used getCurrentFocus in the Notes and ClueList activities.