Initial implementation of history
Requires some shifting so keyboard displays on top of current view. Some
refining is needed: it detracts attention from the action bar, so it
repeats the currently selected clue.

The action bar could be used for the crossword title, and the history
for the clues, but need to make sure first clue is not hidden by keyboard!

Would also be good to have actions on the items (e.g. select the clue).