Some major IO futzing (breaking change IOVersion4)
It's really annoying to maintenance if all changes to IO need to be
reflected in IOVersion1 and IOVersion4, so i rewrote IOVersion4 to put
its new metadata at the end so it doesn't break the chain back to
IOVersion1. Now "default" values for new IO items can be put in
IOVersion1 only.

IOVersion4 was only active for a week, and i updated my personal files.
Apologies to anyone (of whom there are none) who had some IOVersion4
files!