Fix puzzle completion bug
PlayActivity was not calling super onPlayboardChange, and PuzzleActivity
was calling the new dialog even if the timer was not running (i.e. the
puzzle had finished).

It is dubious that the app is using the timer running as a proxy for the
puzzle not being complete: the runTimer pref affects this as well.