Save notes in meta not puz file
Previously unofficial extra across/down notes sections were added. This
is not really respecting the Across Lite format. Instead notes are now
saved in the meta file.

For legacy reasons, notes will still be read from the puz file if found.

This increases the size of the meta file, which may reduce the performance
of the browse activity. We will see...