Almost working implementation of meta cache
Needs longer testing. Instead of reading all meta files (slow on Android
11), store meta data in an internal DB and update with save/load
operations.

If the file changes outside of Forkyz, it will be updated the next time
it is opened in Forkyz, but will be out of date on the browse list until
then. I this this is acceptable as most will not be using multiple apps
on the puzzles.