Fix crash attempting to refresh non-existent puzzle
If the player deleted or archived a file after playing, the next
onResume would still consider that file to be the most recently
accessed, then try to refresh it. Fix by clearing last accessed puzzle
after the first onResume. There's no need to refresh it again unless it
was played again.