Do what i could with hard dpad focus on clue list
Stop calling onKeyUp/Down with dedicated board listener -- it just
causes two calls to the activity method when the key event is not
handled.

Only send left/right to board if it is in focus (so you can change tabs
with hardware dpad). The soft dpad always focuses on the board.

It's clunky, doesn't work that well, but nothing i tried would change
the default behaviour very much... And nothing would make it sensible.