Naturalise zoom behaviour
Touch events need to be reported to both gesture detectors in scrolling
image view. In this case, if a zoom gesture was started with a single
finger down, and the multitouch events are not reported to the gesture
detector, it thinks a long press has occurred.

Remove snapping behaviour on zoom -- neither snap to the top or bottom
of the screen.

Remove timer event on scale -- i'm not sure why it was there, and the
behaviour of highlighting the centre letter was confusing.

Reduce stickiness of screen edges when scrolling, and use fixed sticky
distance (of 1mm) rather than a percentage of board size, which changes
with zoom!

Change double-tap-fit-to-screen behaviour to work as if zoom fit was
selected from the menu rather than the toggle between fit and default
zoom. Toggling back to the default zoom left the board in a different
position to where you started, which makes the old behaviour awkward to
use.

Fixes #24