Add support puzzle source
Each downloader/scraper now has a "support URL" field (which may be
null) where a website for supporting the publication is given.

This is saved in meta data and available via a "Support Puzzle Source"
menu option (that opens the URL) in the play activity.

If there's no explicit "support" link, just a sensible page of the
puzzle source is given instead.