Make puzzleList a list of MutableLiveData
This means we don't refresh the puzzle list each time the meta is
refreshed or puzzles are removed, which in turn means the scroll
position is not lost.

Also avoid a full reload of files on puzzle import, but the adapter is
still refreshed fully, so scroll position is lost.

Needed to add an observer to the SeparatedRecyclerViewAdapter so that
change notifications are properly passed on.