Handle and avoid possible multiple instances
Because the current puzzle/board is stored in the activity instance,
multiple instances (tasks) are not really supported.

First, hack to avoid multiple instances by using broadcasts to ask other
BrowseActivities to close.  All that exist and are not part of the same
task (e.g. not a recreated due to a theme change or rotation), will
close activity (and task on API >= 21). Since BrowseActivity is the
root, this ends all other activities too (i think).

Additionally, make Forkyz robust to the possibility of multiple
instances, so it will behave ok without the above hack.

In particular, since the board/puzzle is stored in the application, it
could be the case that separate instances change the board while the
other is paused. As a workaround, we make the activities avoid the
assumption that the board has not changed since they were paused.

This is not ideal. It occurs because opening a file from the device
triggers a browse activity import, meaning we would have two browse
activities, from which another puzzle can start.

If we make BrowseActivity a singleTask, then the user is always taken
back to the browse activity when launching the app, which is annoying if
you're switching between apps while playing (and singleTask is against
Android UI recommendations).

A longer term fix might be to move the puzzle/board into the/a view
model, then multiple instances can truly exist.